Handle images with digests
Previously, we were splitting image addresses at the _last_ instance of
a colon; this had unfortunate side-effects when said address included a
digest.

This change causes us to look for the _first_ instance of a colon,
allowing both tag and digest to properly co-exist; we don't try to do
anything fancier than that.